+++ apps/api/serializers/api_keys.py
@@ -2,6 +2,7 @@ from datetime import datetime
 from typing import Optional
 from uuid import UUID
 
+from django.core.cache import cache
 from pydantic import computed_field
 
 from apps.api.serializers import Serializer
@@ -20,4 +21,13 @@ class ApiKeySerializer:
 
         @computed_field
         def token(self) -> str:
-            return JWTFactory(str(self.user_id)).api_key(str(self.id))
+            cache_key = f"api_key.{self.id}.token"
+            cached = cache.get(cache_key)
+
+            if cached:
+                return cached
+
+            result = JWTFactory(str(self.user_id)).api_key(str(self.id))
+            cache.set(cache_key, result)
+
+            return result