+++ apps/core/models/acquisition.py
@@ -3,6 +3,7 @@ import hashlib
 from typing import Optional
 
 from django.conf import settings
+from django.core.cache import cache
 from django.db import models
 from django.db.models.signals import post_save
 from django.dispatch import receiver
@@ -59,6 +60,12 @@ class Acquisition(BaseModel):
     @property
     def base64(self) -> Optional[str]:
         if self.content is not None:
+            cache_key = f"acquisition.{self.id}.base64"
+            cached = cache.get(cache_key)
+
+            if cached:
+                return cached
+
             encoded = base64.b64encode(self.content.read()).decode("ascii")
             return f"data:{self.mime};base64,{encoded}"
         return None